+++ changelog.md
@@ -2,6 +2,8 @@
 
 ## Version WIP
 - New option (disabled by default) to show one item per source type in GUI instead of one item per story when showing ALL sources (which is also the start page)
+- Fix source/type reset when redownloading
+- Show the number of images instead of the number of words for images documents 
 
 ## Version 1.6.0